+++ template-lit-js/vite.config.js
@@ -4,21 +4,20 @@ import { crx } from '@crxjs/vite-plugin'
 import manifest from './src/manifest.js'
 
 // https://vitejs.dev/config/
-export default defineConfig(({ mode }) => {
-  return {
-    build: {
-      emptyOutDir: true,
-      outDir: 'build',
-      rollupOptions: {
-        input: {
-          options: path.resolve('options.html'),
-          popup: path.resolve('popup.html'),
-        },
-        output: {
-          chunkFileNames: 'assets/chunk-[hash].js',
-        },
+export default defineConfig({
+  build: {
+    emptyOutDir: true,
+    outDir: 'build',
+    rollupOptions: {
+      input: {
+        options: path.resolve('options.html'),
+        popup: path.resolve('popup.html'),
+        sidepanel: path.resolve('sidepanel.html'),
+      },
+      output: {
+        chunkFileNames: 'assets/chunk-[hash].js',
       },
     },
-    plugin: [crx({ manifest })],
-  }
+  },
+  plugins: [crx({ manifest })],
 })