+++ template-preact-js/src/sidepanel/index.jsx
@@ -1,0 +1,5 @@
+import { render } from 'preact'
+import { Sidepanel } from './Sidepanel'
+import './index.css'
+
+render(<Sidepanel />, document.getElementById('app'))