+++ template-solid-ts/src/sidepanel/index.tsx
@@ -1,0 +1,7 @@
+/* @refresh reload */
+import { render } from 'solid-js/web'
+
+import './index.css'
+import { Sidepanel } from './Sidepanel'
+
+render(() => <Sidepanel />, document.getElementById('app') ?? document.body)